// src/app/models/video.ts
export interface Video {
  id: number;
  title: string;
  user_id: number;
  description: string;
  provider: string;
  source: string;
  namespace: string;
  poster: string;
  duration: number;
  height: number;
  width: number;
  thumbnail: string;
  gif: string;
  updated_at: string;
  created_at: string;
}